* master: (494 commits)
  Adding feedvalidator as a vendorized package, for some day.
  Trying out a new tfidf matcher for subscriptions. Works OK, not great.
  Adding paypal logging.
  iOS v4.6.2.
  Adding saved story tag list when using bottom save story button.
  Fixing shared story feed titling.
  Fixing interactions and activities sizing.
  Fixing landscape ipad keyboard share bug. Stupid typo!
  Adding FAQ entry about story proxy.
  Further removing iphone popover code from ipad.
  Fixing activities to only use URL and title, except Mail which uses the whole body.
  Fixing ipad rotation issues by forcing redraw of background views.
  Fixing view draw issues when selecting interactions and activities from either profile views or the dashboard.
  Fixing issue with Saved Stories crashing app when user has no saved stories.
  Save all/unread/focus selection across app open/close (#633)
  Hide login as menu item if isStaff has changed.
  Fixes to move network calls off main thread and sync feeds for user we have logged in as.
  Initial working login as call.
  v4.6.1 of the iOS app. Submitted to the app store + enterprise deployment.
  Fixing HTML in sent messages.
  ...

+++ apps/reader/models.py
@@ -2,10 +2,14 @@ import datetime
 import time
 import re
 import redis
+from collections import defaultdict
+from operator import itemgetter
+from pprint import pprint
 from utils import log as logging
 from utils import json_functions as json
 from django.db import models, IntegrityError
-from django.db.models import Q
+from django.db.models import Q, F
+from django.db.models import Count
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.cache import cache
@@ -15,6 +19,7 @@ from apps.reader.managers import UserSubscriptionManager
 from apps.rss_feeds.models import Feed, MStory, DuplicateFeed
 from apps.analyzer.models import MClassifierFeed, MClassifierAuthor, MClassifierTag, MClassifierTitle
 from apps.analyzer.models import apply_classifier_titles, apply_classifier_feeds, apply_classifier_authors, apply_classifier_tags
+from apps.analyzer.tfidf import tfidf
 from utils.feed_functions import add_object_to_folder, chunks
 
 class UserSubscription(models.Model):
@@ -80,7 +85,7 @@ class UserSubscription(models.Model):
                     super(UserSubscription, self).save(*args, **kwargs)
                     break
             else:
-                self.delete()
+                if self: self.delete()
     
     @classmethod
     def subs_for_feeds(cls, user_id, feed_ids=None, read_filter="unread"):
@@ -417,8 +422,17 @@ class UserSubscription(models.Model):
         return feeds
     
     @classmethod
-    def recreate_destroyed_feed(cls, new_feed_id, old_feed_id=None, skip=0):
-        user_ids = sorted([int(u) for u in open('users.txt').read().split('\n') if u])
+    def identify_deleted_feed_users(cls, old_feed_id):
+        users = UserSubscriptionFolders.objects.filter(folders__contains="5636682").only('user')
+        user_ids = [usf.user_id for usf in users]
+        f = open('users.txt', 'w')
+        f.write('\n'.join([str(u) for u in user_ids]))
+
+        return user_ids
+
+    @classmethod
+    def recreate_deleted_feed(cls, new_feed_id, old_feed_id=None, skip=0):
+        user_ids = sorted([int(u) for u in open('utils/backups/users.txt').read().split('\n') if u])
         
         count = len(user_ids)
         
@@ -818,8 +832,85 @@ class UserSubscription(models.Model):
             folders.extend(list(orphan_ids))
             usf.folders = json.encode(folders)
             usf.save()
-            
+    
+    @classmethod
+    def verify_feeds_scheduled(cls, user_id):
+        r = redis.Redis(connection_pool=settings.REDIS_FEED_POOL)
+        user = User.objects.get(pk=user_id)
+        subs = cls.objects.filter(user=user)
+        feed_ids = [sub.feed.pk for sub in subs]
 
+        p = r.pipeline()
+        for feed_id in feed_ids:
+            p.zscore('scheduled_updates', feed_id)
+            p.zscore('error_feeds', feed_id)
+        results = p.execute()
+        
+        p = r.pipeline()
+        for feed_id in feed_ids:
+            p.zscore('queued_feeds', feed_id)
+        try:
+            results_queued = p.execute()
+        except:
+            results_queued = map(lambda x: False, range(len(feed_ids)))
+        
+
+        safety_net = []
+        for f, feed_id in enumerate(feed_ids):
+            scheduled_updates = results[f*2]
+            error_feeds = results[f*2+1]
+            queued_feeds = results[f]
+            if not scheduled_updates and not queued_feeds and not error_feeds:
+                safety_net.append(feed_id)
+
+        if not safety_net: return
+
+        logging.user(user, "~FBFound ~FR%s unscheduled feeds~FB, scheduling..." % len(safety_net))
+        for feed_id in safety_net:
+            feed = Feed.get_by_id(feed_id)
+            feed.set_next_scheduled_update()
+
+    @classmethod
+    def count_subscribers_to_other_subscriptions(cls, feed_id):
+        # feeds = defaultdict(int)
+        subscribing_users = cls.objects.filter(feed=feed_id).values('user', 'feed_opens').order_by('-feed_opens')[:25]
+        print "Got subscribing users"
+        subscribing_user_ids = [sub['user'] for sub in subscribing_users]
+        print "Got subscribing user ids"
+        cofeeds = cls.objects.filter(user__in=subscribing_user_ids).values('feed').annotate(
+                                     user_count=Count('user')).order_by('-user_count')[:200]
+        print "Got cofeeds: %s" % len(cofeeds)
+        # feed_subscribers = Feed.objects.filter(pk__in=[f['feed'] for f in cofeeds]).values('pk', 'num_subscribers')
+        # max_local_subscribers = float(max([f['user_count'] for f in cofeeds]))
+        # max_total_subscribers = float(max([f['num_subscribers'] for f in feed_subscribers]))
+        # feed_subscribers = dict([(s['pk'], float(s['num_subscribers'])) for s in feed_subscribers])
+        # pctfeeds = [(f['feed'],
+        #              f['user_count'],
+        #              feed_subscribers[f['feed']],
+        #              f['user_count']/max_total_subscribers,
+        #              f['user_count']/max_local_subscribers,
+        #              max_local_subscribers,
+        #              max_total_subscribers) for f in cofeeds]
+        # print pctfeeds[:5]
+        # orderedpctfeeds = sorted(pctfeeds, key=lambda f: .5*f[3]+.5*f[4], reverse=True)[:8]
+        # pprint([(Feed.get_by_id(o[0]), o[1], o[2], o[3], o[4]) for o in orderedpctfeeds])
+
+        users_by_feeds = {}
+        for feed in [f['feed'] for f in cofeeds]:
+            users_by_feeds[feed] = [u['user'] for u in cls.objects.filter(feed=feed, user__in=subscribing_user_ids).values('user')]
+        print "Got users_by_feeds"
+        
+        table = tfidf()
+        for feed in users_by_feeds.keys():
+            table.addDocument(feed, users_by_feeds[feed])
+        print "Got table"
+        
+        sorted_table = sorted(table.similarities(subscribing_user_ids), key=itemgetter(1), reverse=True)[:8]
+        pprint([(Feed.get_by_id(o[0]), o[1]) for o in sorted_table])
+        
+        return table
+        # return cofeeds
+        
 class RUserStory:
     
     @classmethod
